requirements.txt
flask
dnspython

